+++ Test/Ruleset.ProgramInfo/Test-VirusTotal.ps1
@@ -1,0 +1,94 @@
+
+<#
+MIT License
+
+This file is part of "Windows Firewall Ruleset" project
+Homepage: https://github.com/metablaster/WindowsFirewallRuleset
+
+Copyright (C) 2022 metablaster zebal@protonmail.ch
+
+Permission is hereby granted, free of charge, to any person obtaining a copy
+of this software and associated documentation files (the "Software"), to deal
+in the Software without restriction, including without limitation the rights
+to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+copies of the Software, and to permit persons to whom the Software is
+furnished to do so, subject to the following conditions:
+
+The above copyright notice and this permission notice shall be included in all
+copies or substantial portions of the Software.
+
+THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+SOFTWARE.
+#>
+
+<#
+.SYNOPSIS
+Unit test for Test-VirusTotal
+
+.DESCRIPTION
+Test correctness of Test-VirusTotal function
+
+.PARAMETER Domain
+If specified, only remoting tests against specified computer name are performed
+
+.PARAMETER Force
+If specified, this unit test runs without prompt to allow execute
+
+.EXAMPLE
+PS> Test-VirusTotal
+
+.INPUTS
+None. You cannot pipe objects to Test-VirusTotal.ps1
+
+.OUTPUTS
+None. Test-VirusTotal.ps1 does not generate any output
+
+.NOTES
+None.
+#>
+
+#Requires -Version 5.1
+
+[CmdletBinding()]
+param (
+	[Parameter()]
+	[Alias("ComputerName", "CN")]
+	[string] $Domain = [System.Environment]::MachineName,
+
+	[Parameter()]
+	[switch] $Force
+)
+
+#region Initialization
+. $PSScriptRoot\..\..\Config\ProjectSettings.ps1 $PSCmdlet -Domain $Domain
+. $PSScriptRoot\..\ContextSetup.ps1
+
+Initialize-Project -Strict
+if (!(Approve-Execute -Accept $Accept -Deny $Deny -Force:$Force)) { exit }
+#Endregion
+
+Enter-Test "Test-VirusTotal"
+$File = "C:\Windows\regedit.exe"
+$SigcheckLocation = "C:\tools"
+
+if ($Domain -ne [System.Environment]::MachineName)
+{
+	Start-Test "Remote default"
+	Test-VirusTotal -LiteralPath $File -SigcheckLocation $SigcheckLocation -Domain $Domain
+}
+else
+{
+	Start-Test "Default test"
+	$Result = Test-VirusTotal -LiteralPath $File -SigcheckLocation $SigcheckLocation
+	$Result
+
+	Test-Output $Result -Command Test-VirusTotal
+}
+
+Update-Log
+Exit-Test