+++ Readme/CHANGELOG.md
@@ -58,6 +58,7 @@ If you would like to see fresh changes done since last release you can do so on 
     - `Test-FileSystemPath`
     - `Search-Installation`
     - `Confirm-Installation`
+    - `Test-ExecutableFile`
 
   - `Enable-WinRMServer` and `Disable-WinRMServer` added `-KeepDefault` switch to leave default
   session configurations active, needed to be able to specify `-ComputerName` parameter
@@ -82,6 +83,7 @@ If you would like to see fresh changes done since last release you can do so on 
   - `Set-ATP.ps1` new script used to configure Microsoft Defender Antivirus
   - `Confirm-Firewall.ps1` new experimental script used to validate firewall configuration
   - `Test-RemoteRegistry` new function to test for functioning remote registry
+  - `Test-VirusTotal` new module function used to perform virus total analysis on executable
 
 - Scripts